requirements.txt
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
streamlit==1.25.0
tiktoken==0.5.1
